multiple redir, no more leaks, clean
multiple redirs are working,
I clean the leaks provening from redir
I renamed parser_redir -> parse_redi
