+++ parser/free_ast.c
@@ -6,7 +6,7 @@
 /*   By: qjungo <qjungo@student.42lausanne.ch>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/23 20:01:53 by qjungo            #+#    #+#             */
-/*   Updated: 2023/03/09 15:05:28 by qjungo           ###   ########.fr       */
+/*   Updated: 2023/03/11 11:02:19 by qjungo           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -38,10 +38,6 @@ void	free_ast(t_cmd *cmd)
 	}
 	else if (cmd->type == REDIR_OUT
 		|| cmd->type == REDIR_IN || cmd->type == REDIR_APPEND)
-	{
-		free(cmd->s_redir.filename);
-		cmd->s_redir.filename = NULL;
 		free_ast(cmd->s_redir.cmd);
-	}
 	free(cmd);
 }
