+++ parser/parser.h
@@ -6,7 +6,7 @@
 /*   By: qjungo <qjungo@student.42lausanne.ch>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/23 20:01:50 by qjungo            #+#    #+#             */
-/*   Updated: 2023/03/10 09:59:32 by qjungo           ###   ########.fr       */
+/*   Updated: 2023/03/11 14:52:21 by qjungo           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -45,7 +45,6 @@ typedef struct s_cmd {
 			struct s_cmd	*first_cmd;
 		} s_pipeline;
 		struct {
-			char			*filename;
 			int				fd;
 			struct s_cmd	*cmd;
 		} s_redir;
@@ -55,23 +54,24 @@ typedef struct s_cmd {
 typedef struct s_cmd		t_cmd;
 typedef struct s_minishell	t_minishell;
 
-t_list		*skip_parentheses(t_list *cursor);
+// parser_utils.c
 int			get_token_type(char *token);
-t_bool		are_we_in_parentheses(t_list *tokens);
-t_list		*lst_cut_first_and_last(t_list *tokens);
-int			first_token_redir(t_list **tokens, t_cmd *cmd,
-				t_minishell *minishell);
 
 int			parse_command(t_list *tokens, t_cmd *cmd, t_minishell *minishell);
 int			pipeline(t_list *tokens, t_cmd *cmd, t_minishell *minishell);
 int			set_command(t_list *tokens, t_cmd *cmd, t_minishell *minishell);
 t_cmd		*parser(t_list *tokens, t_minishell *minishell);
 
+// free_ast.c
 void		free_ast(t_cmd *cmd);
 
 int			is_heredoc_valid(t_list *cursor, t_cmd *cmd);
 int			get_heredoc(t_list **token_cursor, t_cmd *cmd);
 
 int			redir(t_list *tokens, t_cmd *cmd, t_minishell *minishell);
+
+// parse_redir_open.c
+int			open_file(t_cmd *cmd, char *filename);
+int			check_error(t_cmd *cmd, char *filename);
 
 #endif /* PARSER_H */
